refactor(desktop): extract command palette data + resolution to commands.ts (F2)
Move the COMMANDS table and the pure command logic (parseCommand,
filterCommands, resolveEnter) out of App.tsx / CommandBar.tsx into commands.ts;
CommandBar and App's executeCommand now call the shared helpers. 9 unit tests
cover aliasing, arg preservation, exact-vs-highlighted Enter resolution, and a
COMMANDS integrity check (no duplicate tokens). The big executeCommand switch
stays as the handler adapter — data-driven dispatch deferred (see plan doc).

No behavior change. Verified: 38 unit tests green, tsc, build, and a Playwright
command-bar smoke (alias run, prefix resolve, :p picker, :theme not auto-actioned).
